Adds starter project for Quartz chapter.
This starter is from the Quartz solution 02_SalesReport. It should
start the student off at the section title Dealing with Pixel
Boundaries (before the Making a PDF section).

This saves the student the work of starting up the project and
adding the boilerplate report objects. It does skip over a
bit of CG drawing which may be worth walking through with the
student.